READ.md
Before you open the app since it uses “pods” the app must be open via the workspace file and not the “xcodeproject” file. Before you run the app, its important to note that this app cannot run on any stimulator, it must be run on a actually Iphone. One reason for this is that this app requires the user’s location and on the stimulator, its settings are already set on a particular location. For example if you was to use it on the stimulator the app would determine your location to be Apple’s headquarters. It really depends on what location you have your Iphone stimulator set too. So when you run the app first thing it does is ask you if its ok to use your location. If the you click no then it will ask you again until you click yes. This is important because without your location the app cannot determine your current location.

 Once you hit yes the app will automatically display the current temperature of your current location. But not only that, it will display the town or city you are currently in on the bottom left. Next to that it will display your current weather conditions. For example if its raining at your location it will say “Rain” or if there’s no significant weather conditions it will say “clear”.  In the middle on the right side it will display a picture of your current weather conditions. So just like before if its raining it will display a picture of rain or if its snowing it will display a picture of snow. 

To the left of the picture it will show you the current temperature of your location in Fahrenheit. Then on top of this display you will see the current date. If you noticed on the top of the screen you would see a search bar. This search bar will give you the ability to search any town or city and find out what the current weather is at that location. For example if you type “London” in the searchbar and click search on the bottom of the keyboard then it would display the weather in London at real time. But remember the search bar only works with cities/ towns so for example if you type in a state name or country name the app will show you results but it will probably be from a city or town that has that name. 

Another cool feature is that it doesn’t matter if the name of a town or city has spaces in the word. So you could type in cites like Las vegas, Fort lauderdale, San francisco and it will work. Also its not limited to the U.S its really any city or town around the world as long as you could type the name using English. If by chance you don’t fully remember a town or city name and the name has two words you could type in the first word and the first letter of the second word and it will display the weather. For example if you type in “Fort L” the app will display the weather for “Fort Laurderdale”. Lastly if you want to look at your weather after looking at other places you can simply type in “Current location” and it will show you current weather again. But remember its case sensitive so it has to be in that exact format.